Added model attributes for ordering

+++ TrainingApp/Exercise+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Exercise+CoreDataProperties.swift
 //  TrainingApp
 //
-//  Created by Bartłomiej Hyży on 26/12/15.
+//  Created by Bartłomiej Hyży on 03/01/16.
 //
 //
 //  Choose "Create NSManagedObject Subclass…" from the Core Data editor menu
@@ -14,6 +14,6 @@ import CoreData
 
 extension Exercise {
 
-    @NSManaged var duration: Int32
+    @NSManaged var duration: Int64
 
 }
